Image Browsing: introduce basic front-end code
This patch introduces some basic structure and scaffolding
for the front-end UI of our first prototype, Image Browsing.

Everything related to this prototype on the front-end should
live within the newly-established `ext.readerExperiments.imageBrowsing`
ResourceModule. Future prototypes should follow a similar convention.
If need be, we can organize our back-end code in a similar way
at some point.

The `ext.readerExperiments.imageBrowsing` module depends on Vue.js,
and will also use Codex at some future point. We will use packageFiles
for this module (and likely for all other prototypes as well). Most
UI code will live in Vue SFCs (which are natively supported in MW).

So far only a single dummy component, "MediaGallery.vue", has been
introduced. We may change the name of this compoent in the future.

On the back-end side, the `Hooks.php` file has been updated to use the
ArticleViewHeaderHook (mostly just to see how this worked). In the
future we might change this back to BeforePageDisplay, or we might
use another hook entirely.

The hook handler adds an empty div to the output page of a given
Article, which is used as the mount point for the Vue.js app.

Finally, the existing PHPUnit test has been removed to avoid
errors in CI; in the future we can re-introduce something else that
is appropriate here.

Bug: T400078
Change-Id: I3f9a40c7aad06a413f3bf1f2f09526bfb8c33c87