Thumbnail extractor from live DOM
This replicates some of the logic from MultimediaViewer,
which lives in mmv.bootstrap.js, for detecting images, excluding
those that live in metadata templates or have opted out of mmv.

Code is run when the stub view component is setup, but just drops
some debug info to console so far:

extractThumbInfo( content ) gives you an array of plain JS objects
which each have references to the live element and surrounding
container, and metadata to be used directly in drawing:
* name
* title (mw.Title object for the File: page)
* extension
* url[width] = { width, height, src }
* alt text if available
* link, container elements if available
* resizeUrl function to render arbitrary-size thumbnails on demand

From here, front-end code can clone the thumbnail images into
the gallery or carousel, including copying alt text directly,
and associate them with their original elements for additional
context extraction or scrolling to the live image.

SVG thumbs that appear in the infobox are excluded since these are
commonly maps or flags that don't blow up well in photos, but photos
in infoboxes are let through. This can be changed easily in the logic.

To match MMV's behavior, any file with an extension not in the default
known list will be excluded as well, which mostly means PDFs and STL
3d objects of things that would've been captured previously. These again
are not expected to fit well in the carousel and would require extra
logic to handle correctly.

Next steps:
* add tests after test runner is merged

Bug: T398992
Change-Id: I0661db07dd89d510a2b17f8045d9d3174815eb09