ImageBrowsing: Add component scaffolding for front-end UI
Adds basic placeholder components / scaffolding for the remainder of
the ImageBrowsing UI. This includes placeholder versions of the
following components:

* Overlay
* DetailView
* VisualTableOfContents
* VisualTableOfContentsItem

The Overlay component lives at the top-level of the app, alongside the
Carousel. When an item in the Carousel is clicked, the "active image"
gets set and the Overlay appears (it uses Vue's built-in Teleport
component to get around some parts of the MediaWiki UI, just like the
Codex Dialog does).

The active image and the set of all images are passed to the Overlay
component as props; Overlay in turn passes these props to the
DetailsView and VisualTableOfContents components, respectively.
Clicking on a different item in the VisualTableOfContents could change
the active image by emitting an event.  Clicking a close button clears
the active image in the same way. Once the active image is cleared,
the Overlay component becomes hidden automatically, but it will appear
again if a new Carousel Item is clicked.

Styling is extremely crude for now with the expectation that follow-up
commits that focus on specific components can work to match the designs;
this patch is just to help get the basic structure established quickly.

Even though this is a JS codebase, the ThumbnailImageData interface in
"types.d.ts" is used (via JSDoc annotations) wherever possible. Having a
clearly-defind interface that is used across various components should
hopefully simplify further development.

Change-Id: I87e71c794063d44e6d96ae13324e24b030dfc748