Visual ToC initial implementation
Includes thumbExtractor updates to properly handle lazy-loaded images,
and cleans up the url extraction which is not needed for now.

Image visual ToC/scroll view showing other images from elsewhere
in the article and their caption or other context to jump around
the article view in.

Uses the caption from the figcaption if present, or alt or image name
otherwise; attempts to show the image at a modest size, while not yet
styling the positioning of the ToC as that needs to be hooked onto the
detail view.

Currently using plaintext version of caption; can switch to HTML
to support internal links if desired.

Clicking on an image selects it in the carousel, though this doesn't
really look great yet :D

Clicking the 'view in article' link closes the overlay and jumps to the
image in the view, seems to work but not carefully tuned yet.

I noticed that propagating events to the higher level widgets is a pain,
possibly we should be using a shared event bus to emit higher level
events to like 'select image', 'view image in article' etc.

Bug: T401221
Change-Id: Ie3cc6d3d4e7c05dccdfab484de05e99827663c31