ImageBrowsing: Introduce "useBackgroundColor" composable
The DetailView component is quite large and complex, so we should break
things out into separate concerns where possible to simplify further
devleopment.

One way to do that in Vue.js is through "composables" – re-usable
pieces of functionality that can be dropped in to the setup()
function for components using the Composition API. More on Vue.js
composables can be found here:
https://vuejs.org/guide/reusability/composables.html#composables

This patch introduces a new "useBackgroundColor" composable that wraps
the FastAverageColor library. Now the consuming component (in this case
the DetailView) can remain ignorant about the details of how the color
is extracted. Instead, just provide the width, height, and src of the
desired image and get a FastAverageColorResult object in response. That
object exposes a number of properties that can be queried to get hex or
rgb(a) values, whether the color is "light" or "dark", etc.

Note: To get the color, we need to actually decode the image; this is
an async operation. To simplify the code I've opted to use async/await
for this (this is finally OK in MediaWiki JS code). One consequence of
this is that the consuming component for our async composable (the
DetailsView component) nees to be wrapped in Vue's built-in Suspense
component to define the "suspense boundary" – where the async part of
the UI begins basically. For more on Suspense, see:
https://vuejs.org/guide/built-ins/suspense.html#suspense

In the future the use of Suspense here could allow us to easily define
a loading state in the #fallback slot of the Suspense component.

Bug: T401121
Change-Id: I8a03b6a9e0afb43d4b852f8ffc6c4cf2e419ced3