ImageBrowsing: Use template ref for Detail view caption
The Detail View code currently performs a check for overflowing
caption text wrapped in a "nextTick" callback.

Vue.js actually provides a better way to do solve this exact
problem: needing to access some property of a DOM element which
has not yet been rendered at the time of the setup function:
template refs.

https://vuejs.org/guide/essentials/template-refs.html#template-refs

The caption text element already has a "ref" attribute, so we
can just pass that to Vue's built-in "useTemplateRef" composable
to get a handle for our yet-to-be-rendered element that we can
use later.

Then, instead of manually handling the next tick, we can just
add an "onMounted" handler to perform any DOM-specific calculations
once the page has been loaded.

Bug: T401121
Change-Id: If6308987511431f95920288962b2b5aa23efad6a