Use codex variables for some colors in Overlay, VTOC
This fixes the background of the overlay in Dark Mode and
uses proper colors (and hover colors) for the 'view in article'
link on Visual TOC entries.

Also Less-ifies the syntax making it more readable, and tidies
up some of the padding and sizing to match the mockups from
design better.

Now using proper Vue component for Codex button, so should have
the standard behavior now.

Cleaned up styles to match a little better and for now using the
full column on wider-than-320px phones.

Set 'view in article' scrolling to use smooth behavior, which
looks much nicer!

Follow-up on earlier work for T401221

Bug: T401221
Change-Id: Ie5b4794fb2c74ada23c1ea431e62b83244faa3a6