Image Browsing: VTOC masonry layout
Introduces a basic "pinterest" / masonry-sytle layout
with 2 columns for the larger-screen layout of the
VisualTableOfContents.

The parent VTOC component applies a standard CSS grid
layout, with 2 flexible columns. grid-auto-rows is used
to define a repeating series of 10px-high rows where the
child elements (VTOC-item components) will be placed.

By itself, this would result in a jumbled layout where
all the grid items appear stacked on top of one another
staggered by 10px. We need to calculate the real, rendered
height of each VTOC item (something we cannot know until
the components have mounted).

In order to do this, the VTOCItem component has been updated
to use the "useResizeObserver" composable from Codex. This
returns a constantly-updated set of width and height values
for any element whose template ref is provided as an
argument. That computed height value is then rounded to
the nearest 10px (corresponding to our grid-auto-rows in
the parent). Finally, a computed string is generated
using the template: `span ${ computedHeight.value / 10 }`.
This allows for dynamic binding of the 'grid-row-end'
property to this value on the VTOCItem root element,
resulting in inline style attributes that look like this:

    style: "grid-row-end: span 42;"

Where the numerical span value corresponds to the height
of the given VTOCItem rounded to the nearest 10px.

Bug: T402964
Change-Id: I1209c0439f29ab75072ee0016aec924424ee5309