Fix scroll-to-top on VTOC selection
Now correctly scrolls back to the top of the overlay when tapping
on an image in the visual TOC / scroll view, showing you the zoomed
image.

Much better!

This scroll "within" the overlay lives in the Overlay component, and it
does _slightly_ reach into the DetailView component to get its $el
in order to scroll to it but this should be a very loose coupling.

This should allow flexibility without exposing raw HTML details across
components.

Bug: T401221
Change-Id: I4c691a10277f33dcf55bc650555577870b8dfe32