ImageBrowsing: primary metric instrumentation
The primary metric is the image carousel's click-through rate, i.e.,
the ratio of unique users who clicked on an image in the carousel
to the amount of unique users who are shown the carousel.

Instrumentation is implemented in 3 places:
1. on image carousel load
2. on image click
3. on detail view close

Note that the `action_context: 'image1'` interaction data is added on
the first image click.

See instrumentation spec at
https://docs.google.com/spreadsheets/d/1ZqmKZm7-hdZ1HPDx8KSgZPUaJx7l-Wcd6i7sM5ZpfNI/edit?gid=0#gid=0

Bug: T403255
Change-Id: Idc0a29613a5fa7732b9689b38f272b5b333a7c7b