ImageBrowsing: DetailView caption updates after selection
Ensure the caption text updates when a new image is selected
from the VTOC. Before the composable received `props.image` as
a static value, not a reactive reference. Now when a new image
from the VTOC is clicked, the caption text updates becauase the
composable reactively updates and watches for changes to the
image prop. The composable param type is now an ImageData object
or a ImageData reactive reference.

Bug: T402223

Change-Id: Idf1f8b7863c5398946a8590d6dcb2d9c11f41ee8