ImageBrowsing: Ensure the "more" button updates along with caption
Now that reactive caption updates are working properly, we also
need to update the DetailViewCaption component to re-calculate
whether or not the "more" button is necessary (based on whether
there is overflowing text).

This change re-organizes some relevant code for clarity and adds
a watcher that re-runs the original check performed when the
component mounts any time the caption updates.

Bug: T402223
Change-Id: Ib91d329082ef49f03feccebf4e1b9159535035df