ImageBrowsing: fix DetailViewControls reactivity
When an image is selected in the VisualTableOfCntents (VTOC), the
image info should be reactive and update in the DetailView. The
problem is that the `decriptionUrl` computed property depends on
the image info fetched once during the component setup and doesn't
update when the image prop changes.

Now, the `descriptionUrl` automatically updates to reflect the
new image's description URL. This ensures the share popover shows
the correct URL and the "View on Commons" link navigates to the
correct page.

* Track image prop changes with a watcher. This watcher refetches
image info when the image prop changes.
* Change `imageInfo` from a computed property to reactive reference.
* Add `fetchImageInfo()` function. This is called initially when the
component mounts and when the image prop changes.
* View on Commons link now opens a new tab the correct URL.

Bug: T404228
Change-Id: I16b33b876e1d5df3d3b8c4e2d67a47c096386e9b