ImageBrowsing: Use a Vue Plugin for instrumentation
This patch moves the existing instrumentation code into a
new dedicated Vue.js plugin, plugins/instrumentationPlugin.js.

Usage is similar to what existed before, but some logic has
been added to the initialization code to perform a runtime
check for the "ext.eventLogging" ResourceModule. The presence
or absence of that runtime "soft dependency" will alter the
configuration passed to the instrumentation plugin; if the
event logging extension is not found then all instrumentation
for our application is disabled.

The plugin itself uses the provide/inject API to globally
provide a submitInteraction method (with some arguments
pre-populated by the plugin config) to all components in the
app tree.

If the instrumentation is not enabled, then calling this method
in component code (which still happens) amounts to a no-op. But
if instrumentation is enabled, then mw.eventLogging.submitInteraction
is called with the appropriate arguments.

For more information about Vue plugins, see:
https://vuejs.org/guide/reusability/plugins

Bug: T403259
Change-Id: Iff6ae86f18b94639e1c985eb9c336fc030d59de1