Add router/history navigation
Instead of directly updating the activeImage when
a (new) image gets selected, we'll add a routing layer
inbetween. That'll update the browser history to record
the navigation, which the router will then listen to and
update the activeImage.

Changes in history, or links that were shared, will also
get picked up be the router and invoke the same change
to activeImage, which will trigger the rest of the app
to load up the relevant image.

Bug: T404397
Change-Id: I8d981b532f18c9f4718c989e2e5ce955d62f8e57