Remove unnecessary high-resolution image download
We were previously fetching the original image file in full resolution
for calculating the crop position; this wasn't necessary, and massively
slowed down loading time on slow networks.

Switched it to using the roughly screen-sized thumbnail we already
fetched, which now uses a crossorigin=anonymous marker so we can
directly use it to read pixels.

This should be a simpler, more conservative implementation of roughly
the main optimization from Matthias's earlier patch at:
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/ReaderExperiments/+/1186624

May want to double-check that coordinates are all correct, this is a first
draft and not heavily tested yet.

Bug: T404378
Change-Id: I4a70033a2db7ae86e96bc4c3ac09552af9dd9b34