Precalculate average colors from thumbnails
This in most cases eliminates the flash of base/neutral bg
color behind the detail view when loading it, by precalculating
the average colors as each thumbnail in the carousel or vtoc
loads.

If the carousel or vtoc thumbs don't load in time, the
behavior should fall back to using the neutral background color
when switching to that image until the detail view is loaded,
at which point it will be used to calculate the average color.

* carousel and vtoc images now use crossorigin=anonymous
  for CORS compatibility so we can read them directly
* keeping a Map of image.name->ColorResult as a handy
  cache of completed color calculations
* useBackgroundColor now takes a second parameter with an
  image reference, which will update the result with the
  calculated color when the first thumbnail loads for
  that image
* the reactive references "shake this out" as before; the
  existing useBackgroundColor() return value is the same
* interrupting a download by clicking away fails gracefully
* worked around state update issue in Firefox

Bug: T405528
Change-Id: I034b30e8847c0a21e713875bd6b7a346a7b23547