Lazy-load carousel and visual TOC items
Use the browser-native lazy-load feature (`loading="lazy"` attribute)
to spread out download cost of thumbnails in the carousel and visual TOC.

Previously these would cause floods of thumb downloads on initial page
load (carousel) and overla opening (visual TOC) even if nobody ever
scrolled to view the additional images.

On a slow 3G or 4G network this was a large cost and could delay loading
of the detail view.

Also added width and height elements on the visual TOC images, which
allows the browser layout engine to correctly lay out the entire overlay
even before the images start loading.

This does mean that you're more likely to catch an unloaded or in-progress
image while scrolling in the carousel or VTOC; currently these use the
standard browser behavior of rasterizing in during load and will show over
a light gray or dark gray background depending on light/dark mode (it's
the standard button background, which wraps the image).

I could not get the previous styles using max-width and max-height to
correctly reserve space until loading actually began, so have changed
the styles to width/height/object-fit. This causes a small visible change
with very wide images, which are now "letterboxed" in the button background
color (the same background color that appears in the bounding box before
and during load).

Bug: T405523
Change-Id: I9ddde43857a73bc621f4dd036b854ec45b1d6eaa