Dependencies: Override Glob to use supported version
Grunt (which we use as a task runner for front-end linters)
depends on an old version of the Glob package, which in turn
depends on a vulnerable and unmaintained package called inflight.

This patch uses NPM's "override" feature to force all packages
in this project to use Glob 10.x or newer, which is supported
and has removed the vulnerable dependency.

Behavior for scripts like "npm test" seems unchanged.

Bug: T405993
Change-Id: I10432da47445102f90e1717906615ef1fe39956c