Fix compatibility with iOS 16, 17
We were using RegExp.escape which is a great feature but it's
very new, and isn't available on older iOS versions.

Since we've got a fixed string there's really no need to call the
function anyway; replaced it here with a literal regexp that's
defined right alongside the string prefix so they're easy to keep
in sync.

Tested with full functionality on iOS 16, fixing the most
important regression (iPhones still on iOS 16 or 17).

Not addressed here:

In my local testing, iOS 15 and iOS 12 have some troubles with
CORS requests, this may be a local config issue and I have not
tested with the live wiki config.

iOS 12 seems to also have some style setup problems with the
detail view, which are out of scope for this task.

Bug: T406365
Change-Id: I1dcb493d7782c44fadca11aff48d8ed741f97168