Fix fetching media from other wikis
The original code had some issues:
- it assumed that the order of the results matched
  the results relevant (which is not the case when
  used as generator; hence the addition of the
  "index" parameter), which caused repeat (and
  missed) results since the "offset" for a follow-up
  call didn't necessarily map to the point in the
  results I thought I was at...
- it assumed globalusage data would be available in
  order of results, which is also not true - no
  globalusage data in an earlier result may not mean
  that it's a bad result, but that we don't yet have
  the data
- it also didn't consider that imageinfo maxes out
  at 50, which was something that could happen for
  well-illustrated articles already

This should all be fixed through proper use of the
continuation params to paginate the results, along
with some optimizations to attempt to progress the
offset as early as possible to reduce response
payloads. This has become too convoluted by now, but
AFAICT, it works.

Also swapped out the classic search profile for
'popular_inclinks', which seems to be a closer match
to the kind of results we want to provide here:
those reused more extensively.

Added a test case that runs useMediaSearchResults
with a mocked API wrapper that contains actual lookups
from the 'London' article and the expected combined
result JSON. This should guard against regressions
if we poke this code later.

Bug: T406139
Change-Id: I3b7ff6268675de0e7e574f36054addb4f4a4f725