ImageBrowsing: add legacy browser support
* Resize Observer is unsupported in legacy browsers
so now we check whether it is supported and apply
custom VTOC grid styles. If it's supported, then apply
the `grid-row-end` style for "masonry" tile grid. If
it's not supported, then apply a simpler custom grid
style. Before, there was an issue with calculating
the height, which uses ResizeObserver and resulted in
`grid-row-end: span NaN` in unsupported browsers.
* Simplified the caption link styles by removing the
`where()` CSS pseudo-selector. It's unsupported in
legacy browsers, which caused undesirable styles. It
was unnecessary to use only one selector inside it.
* Fix the DetailView height. Modern browsers use
`dvh` while legacy browsers use `vh`. This should
fix the DetailView control buttons from overlapping
each other.
* Ensure the caption text is aligned at the end of
the flex container by replacing `end` with
`flex-end`. After resolving the DetailView height
for legacy browsers (eg. iOS 12), the caption text
was incorrectly displayed at the top, instead of
the bottom.

Bug: T406379
Change-Id: I43d6ce666d6400ecd90c4c5c65465a2791c31aef