packageFiles: Move init.js to front of list
For packageFiles to work the first file is typically the entry
point. I think this was working because while it is second in the
list it is the first JS file but it would be less confusing to
make this more explict. Also in future this could lead to errords if
the ordering every changes.

Change-Id: I84dab741300ae8a8d453a4474e26483147dbea5c