Fix potentially broken cache in useContentImages
useContentImages was keeping a cache of previously extracted images
in an object, with the key indexed by the element being searched.

Since object property keys are either Symbols or flattened to strings,
this would result in keys being in the form "[object HTMLDivElement]"
meaning if we checked different divs they would overwrite each others'
values.

Changed from an object to a standard ES5 WeakMap where keying on
an element works. Using a WeakMap instead of a Map is standard
practice for caches, as it allows memory to be freed if the element
is destroyed, though that's unlikely to make a difference here.

In practice this makes no difference at present as we only ever
check one element, but now it's correct. :D

Change-Id: Iae6a1443b609c3b0f28e7769a28788e9ffa3c0f6