WIP: ImageBrowsing: fix background and canvas calculations on iOS 12
iOS 12 Safari seems to be very picky about the order of
attributes on an <img> element and needs the `crossorigin`
to appear before `src` or the data taints the canvases.

This may be related to how vue creates HTML based on templates.

Newer browsers (confirmed iOS 15 Safari and later) seem
to handle the previous markup fine but iOS 12 would consider
all the images tainted so the background colors came out
failed even after we fixed redirect issues.

Unresolved issues:
* This change introduces a new bug in iOS 12, which results in
  the infobox images not loading in the Carousel and DetailView.
  Initial testing indicates this may be cache-related in Safari,
  and likely difficult to fix reliably.

Bug: T406379
Change-Id: I70679bd02c3ecb5cb8f7d389a715a8f17bc7d412