Cleaner load experience
This change does 2 things:

- turn the carousel images into croppedimage instances,
  so they gain the loading "animation"
- set a fixed height server-side to prevent FOUC

The former speaks for itself, but the latter is doesn't
"actually" solve the FOUC problem - it just moves the FOUC
around: instead of FOUC on pages with the carousel, we'll
now have one on pages that don't have the carousel.

IMO, this is better because:
- I have a strong suspicion that, in terms of pageviews,
  the <has carousel> case is much more prominent than the
  other
- Gently collapsing down superfluous whitespace feels like
  a less jarring experience than content being pushed away
- In terms of the experimental nature of this project, the
  case where we don't have a carousel does not matter, and
  even if it's a bad experience, it will not affect our
  metrics

This should eventually get a proper fix, but said proper
fix is overkill/too late in this late stage of the project.

Bug: T405382
Change-Id: Ic97b0b8ecbd798c856db9876db73e10197bedd33