StickyHeaders: Add new resource modules
Introduces a series of new resource modules for the StickyHeaders
prototype. The are:

- ext.readerExperiments.stickyHeaders (targets parsoid)
- ext.readerExperiments.stickyHeaders.legacy (targets legacy parser)
- ext.readerExperiments.stickyHeaders.styles (common styles for both)

The discrepancies between markup in Parsoid vs legacy parser outputs
are enough that splitting the feature code into two entirely separtate
modules seemed like the best approach.

The hooks.php file performs a check (identical to what is done in
MobileFrontend) to determine whether or not the user is on Parsoid;
the appropriate module is loaded based on this check.

Regardless of parsoid/legacy rendering, we are relying on a new
dedicated JS hook called "readerExperiments.section-toggled" which
is fired from MobileFrontend code (in both legacy and parsoid
code-paths). This patch depends on a patch in that project to add
these new hooks.

Bug: T408046
Change-Id: I2460c32539ca00032767f7ff7ce89a2bef089687
Depends-On: I1a95522c422c3c4dc24a66a832aff9bf9fbe72c0