Remove broken thumb load optimization
Removes a bit from d5330930 which tries to add a loader handler
to the original thumbnail to use it for calculations if available;
turns out this doesn't actually work as the thumbnails are not
set up as cross-origin so don't have the right header exchange
for permission to use them.

This was harmlessly hidden by exception handling, except on
Safari where the runtime itself sends an error to the console
with a giant incomprehensible backtrace, on iOS 15 and on current
macOS.

Removing this code should behave better and avoid the error log
spam without having any performance impact on production.

Bug: T406379
Change-Id: I03d4b8881949681c05f823f8cefdbd7016d4e1fc