StickyHeaders: scroll-margin-top fixes
Legacy parser was applying incorrect scroll positions on headings
in sections from MobileFrontend's `Toggler.js` while the page was
open and active. Both legacy and parsoid produced incorrect positions
on initial page load when StickyHeaders was active as the `position:`
`sticky` bit would obscure the target elements for h3 and smaller.

For now, rather than alter MobileFrontend we're overriding these
incorrect scrolls with correct scrolls using `element.scrollIntoView()`
which takes `scroll-margin` and `scroll-padding` and such into account.

The original scroll attemps and the correct scrolls happen during the
same event loop run, and the intermediate position is never seen on
screen.

This fixes links to h3 sections and smaller, and to any `div`, `span`
etc with an `id` attribute used for manual fragment links, when
StickyHeaders is active.

Bug: T409349
Change-Id: I8ed19ba66b96709e1664794a0a05ee67a0654529