Move away sticky headers for content no longer in viewport
This adds an intersection listener to only add the sticky
class to headers for which the content are within viewport.
Next, it adds a scroll listener to go over those headers
(in view) to determine the position of their content and
make sure they don't extend beyond it (where they'd overlap
with the next section)
This effectively moves previous sticky headers out of the
way as a new one is scrolled to the top.

I looked into alternative without scroll listener, but
found none. There is no good timing for removing the sticky
header (removing it before it extends beyond the next may
be too soon when the next heading is smaller in height;
i.e. has not yet reached top:0 position), and tinkering
with stacking order (z-index etc) gets messy, and would
require adding a background to the sections (so it fully
covers up lingering headings)

All in all, this "scroll out of the way as a new heading
is supposed to take its place" is also the most visually
satisfying solution as well, IMO. I would've prefered a
fix that didn't involve a scroll listener, but at least
it's a pretty light one - throttled, and only operating
on elements within the viewport.

Bug: T410441
Change-Id: I0505b0965598edd1867267871718d48a3c63b6bb