StickyHeaders: reworked fix for ul, ol bullet points "peeking"
Instead of targeting the specific elements to keep them from
going outside the margins, just add our own background in each
heading that extends beyond the margins with the opaque BG.

Works for both legacy and parsoid, and should handle any future
cases of "margin spillage" that aren't z-index related.

Note I'm not 100% sure why I needed the z-index: -1 on the
sticky-background div, it should be behind everything already
but the heading was showing up behind the background without
it.

Bug: T409325
Change-Id: Ib73a107225d7a86a4c92d39a984a1302984317ff