StickyHeaders: possible fix for 'bouncing' on sticky offset
Could occasionally reproduce a state where it would change
sticky state on a header at high frequency when scrolling
to a particular position.

Have only been able to reproduce it on beta so far, uncertain
if this resolves it.

Redid the state management of `intersectingHeadings` from
an array that could grow in definitely to a Set that's
managed more carefully.

Also changed the pre-decided offscreen check from "> 0"
to ">= 0" just to make sure no stray fractional pixels
spill over.

And teaked the state management to remove the CSS bits and
the state Set removal at the same time, seems clearer.

...and now also making sure we update the 'top' when scrolling
fast and new things come into view.

Bug: T409325
Change-Id: I8a5a938703f875fe2877beeb1fc563865e32951e